infobi-2025/apps/backend/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# Database - Multi-DB Support
sqlalchemy>=2.0.0
pyodbc>=5.0.1
aioodbc>=0.5.0
pymysql>=1.1.0
psycopg2-binary>=2.9.9

# Apache Arrow
pyarrow>=14.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# Excel Export
openpyxl>=3.1.2

# Utility
aiosqlite>=0.19.0
jinja2>=3.1.3